Moved default files so that they will be processed by Babel
This is based on the rule that adds an exception for the app folder in package-webpack.
