+++ packages/roc-package-web-app-react/app/default/client.js
@@ -1,4 +1,4 @@
-import { createClient } from '../app/client';
+import { createClient } from '../client';
 import getRoutesAndStore from './get-routes-and-store';
 
 const { store, routes } = getRoutesAndStore(true);
