+++ packages/roc-package-web-app-react/package.json
@@ -14,7 +14,6 @@
     "lib",
     "styles",
     "views",
-    "default",
     "bin",
     "app"
   ],
